ACMS-264: Some cleanup and refactoring
