+++ modules/acquia_cms_search/acquia_cms_search.libraries.yml
@@ -1,5 +1,5 @@
-facets:
+page:
   js:
-    js/search-facets.js: {}
+    js/search-page.js: {}
   dependencies:
     - core/jquery
